dev-irq: provide two different DMA buffers
A DMA needs a memory to write the data to; this can either be separate
memory (e.g. DDR4 attached to Programmable Logic (PL ~ FPGA)) in Zynq
which is then accessed over PCIe or - if the program already runs on
ARM - some part of the PS memory, set aside as a DMA buffer.